# api/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Backend API",
        "description": "BillPays API RESTful",
        "contact": {
            "name": "Vitor Rodriuues",
            "url": "https://github.com/vs0uz4/cr_laravel_com_android",
            "email": "vitor.rodrigues@gmail.com"
        },
        "version": "0.0.1"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/bill_pays": {
            "get": {
                "tags": [
                    "BillPays"
                ],
                "summary": "Display a Listing of the BillPays",
                "description": "Display a Listing of the BillPays",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return a Collection of BillPays"
                    }
                }
            },
            "post": {
                "tags": [
                    "BillPays"
                ],
                "summary": "Create an BillPay",
                "description": "Create an BillPay",
                "operationId": "store",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "date_due": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "value": {
                                    "type": "number"
                                },
                                "category_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Return Created BillPay"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/bill_pays/{id}": {
            "get": {
                "tags": [
                    "BillPays"
                ],
                "summary": "Display an BillPay",
                "description": "Display an BillPay",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the Found BillPay"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "BillPays"
                ],
                "summary": "Update BillPay",
                "description": "Update BillPay",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                },
                                "date_due": {
                                    "type": "string",
                                    "format": "date"
                                },
                                "value": {
                                    "type": "number"
                                },
                                "category_id": {
                                    "type": "integer"
                                },
                                "done": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the Updated BillPay"
                    },
                    "404": {
                        "description": "Resource not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "BillPays"
                ],
                "summary": "Delete an BillPay",
                "description": "Delete an BillPay",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "500": {
                        "description": "Resource can not be deleted"
                    },
                    "404": {
                        "description": "Resource not Found"
                    }
                }
            }
        },
        "/bill_pays/total": {
            "get": {
                "tags": [
                    "BillPays"
                ],
                "summary": "Calculate totals and counts of BillPays",
                "description": "Calculate totals and counts of BillPays",
                "operationId": "calculateTotal",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return array of Totals and Counts of BillPays"
                    }
                }
            }
        },
...
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Display an Category",
                "description": "Display an Category",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the Found Category"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update Category",
                "description": "Update Category",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the Updated Category"
                    },
                    "404": {
                        "description": "Resource not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete an Category",
                "description": "Delete an Category",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer __token__",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "500": {
                        "description": "Resource can not be deleted"
                    },
                    "404": {
                        "description": "Resource not Found"
                    }
                }
            }
        },
...
    },
...
}